+++ services/advisor/mgmt/2017-04-19/advisor/CHANGELOG.md
@@ -1,7 +1,30 @@
-# Change History
+# Unreleased
+
+## Breaking Changes
+
+### Struct Changes
+
+#### Removed Structs
+
+1. SetObject
+
+### Signature Changes
+
+#### Funcs
+
+1. RecommendationMetadataClient.Get
+	- Returns
+		- From: SetObject, error
+		- To: MetadataEntity, error
+1. RecommendationMetadataClient.GetResponder
+	- Returns
+		- From: SetObject, error
+		- To: MetadataEntity, error
 
 ## Additive Changes
 
-### New Funcs
+### Struct Changes
 
-1. Resource.MarshalJSON() ([]byte, error)
+#### New Struct Fields
+
+1. MetadataEntity.autorest.Response